try to test

+++ server/db/tags.go
@@ -1,11 +1,15 @@
 package db
 
-//easyjson:json -snake_case
+import (
+	"fmt"
+	"log"
+)
+
 type Tag struct {
 	ID uint `json:"id"`
 
 	UserID uint   `json:"user_id"`
-	Name   string `json:"name"`
+	Name   string `json:"tag_name"`
 	Color  string `json:"color"`
 }
 
@@ -14,7 +18,7 @@ type Tags []Tag
 func (u Tag) GetAll(userID string) (Tags, error) {
 
 	rows, err := DB.Query(
-		"SELECT id, name, color FROM tags WHERE user_id = $1",
+		"SELECT id, tag_name, color FROM tags WHERE user_id = $1",
 		userID,
 	)
 
@@ -22,13 +26,12 @@ func (u Tag) GetAll(userID string) (Tags, error) {
 		return nil, err
 	}
 
-	var tags Tags
+	tags := make(Tags, 0)
 	var tag Tag
 
 	for rows.Next() {
 		err = rows.Scan(
 			&tag.ID,
-			&tag.UserID,
 			&tag.Name,
 			&tag.Color,
 		)
@@ -47,22 +50,26 @@ func (u Tag) GetAll(userID string) (Tags, error) {
 }
 
 func (tag Tag) Create() error {
+	fmt.Println("TAG", tag)
+	fmt.Println("DB ", DB)
 	_, err := DB.Query(
-		"INSERT INTO tags (user_id, name, color) VALUES ($1, $2, $3)",
-		tag.UserID, tag.Name, tag.Color,
+		"INSERT INTO tags (user_id, tag_name, color) VALUES ($1, $2, $3)",
+		&tag.UserID, &tag.Name, &tag.Color,
 	)
 	if err != nil {
+		log.Println(err)
 		return err
 	}
 	return nil
 }
 
-func (tag Tag) Update() error {
+func (tag Tag) Update(userID string) error {
 	_, err := DB.Query(
-		"UPDATE tags SET (name = $2, color = $3) WHERE id = $1",
-		tag.ID,
-		tag.Name,
-		tag.Color,
+		"UPDATE tags SET (tag_name = $2, color = $3) WHERE id = $1, user_id = $2",
+		&tag.ID,
+		&tag.Name,
+		&tag.Color,
+		&userID,
 	)
 
 	if err != nil {
@@ -72,10 +79,10 @@ func (tag Tag) Update() error {
 	return nil
 }
 
-func (tag Tag) Remove(id string) error {
+func (tag Tag) Remove(userID string) error {
 	_, err := DB.Query(
-		"DELETE FROM tags WHERE id = $1",
-		id,
+		"DELETE FROM tags WHERE id = $1, user_id = $2",
+		tag.ID, userID,
 	)
 	if err != nil {
 		return err
